requirements.txt
streamlit==1.33.0

# Core LangChain ecosystem (compatible set)
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52

# Groq integration
langchain-groq==0.1.3

# Embeddings
sentence-transformers==2.6.1
transformers==4.40.2
torch==2.2.2

# Vector DB (Astra / Cassandra)
cassio==0.1.7
cassandra-driver==3.29.1

# PDF processing
PyPDF2==3.0.1

# Utils
python-dotenv==1.0.1